do batch appends in readRDDFromFile instead of all at once
readRDDFromFile attempts to read a whole file and create a RDD when the file is completely read. This limits the usable size of an input file  to the size of client side memory. It'd be better if it was created in finite sized chunks and added to.

When I tested this patch it was very slow when using the default batchSize of 1024. With larger batch sizes it seemed fine.

Should the intermediate RDD be cached and/or checkpointed?